refactor(core): move mocks to core/perspective
