+++ packages/sanity/src/core/releases/hooks/__tests__/useVersionOperations.test.tsx
@@ -2,9 +2,9 @@ import {act, renderHook} from '@testing-library/react'
 import {beforeEach, describe, expect, it, vi} from 'vitest'
 
 import {createTestProvider} from '../../../../../test/testUtils/TestProvider'
+import {usePerspectiveMockReturn} from '../../../perspective/__mocks__/usePerspective.mock'
 import {useReleaseOperationsMockReturn} from '../../store/__tests__/__mocks/useReleaseOperations.mock'
 import {useVersionOperations} from '../useVersionOperations'
-import {usePerspectiveMockReturn} from './__mocks__/usePerspective.mock'
 
 vi.mock('../../store/useReleaseOperations', () => ({
   useReleaseOperations: vi.fn(() => useReleaseOperationsMockReturn),
