+++ packages/sanity/src/core/releases/tool/overview/__tests__/columnCells/ReleaseName.test.tsx
@@ -3,11 +3,11 @@ import {describe, expect, it, vi} from 'vitest'
 
 import {mockUseRouterReturn} from '../../../../../../../test/mocks/useRouter.mock'
 import {createTestProvider} from '../../../../../../../test/testUtils/TestProvider'
-import {activeASAPRelease, archivedScheduledRelease} from '../../../../__fixtures__/release.fixture'
 import {
   mockUsePerspective,
   usePerspectiveMockReturn,
-} from '../../../../hooks/__tests__/__mocks__/usePerspective.mock'
+} from '../../../../../perspective/__mocks__/usePerspective.mock'
+import {activeASAPRelease, archivedScheduledRelease} from '../../../../__fixtures__/release.fixture'
 import {releasesUsEnglishLocaleBundle} from '../../../../i18n'
 import {type InjectedTableProps} from '../../../components/Table/types'
 import {ReleaseNameCell} from '../../columnCells/ReleaseName'
